requirements.txt
requests
pyyaml
datasets<3.0
pynvml
mlx-lm
numpy
scipy